feat: added AbstractFIFO and CircularBuffer classes
Precommit-Verified: 1cc0ee5926f9072be53c386c759b27d3d1b1671976570ed0ea371cc1edfcdf01

+++ libs/limes_core/threading/Threading.cpp
@@ -24,7 +24,7 @@ int maxNumThreads (int defaultVal) noexcept
 		const auto count = CPU_COUNT (&cpuset);
 
 		if (count > 0)
-			return count;
+			return static_cast<int> (count);
 	}
 #endif
